Fix for tag race condition and disabling pushing of latest
If there are multiple agents sharing the same machine and are building an image for the same repo in the same registry, then they would currently encounter a race condition as they'd tag their images with the same name. This commit attempts to solve the issue by introducing the buildkite commit sha to the original name. We also enabled the ability to skip deploying to production by checking to see if an envar for disabling the push of latest is set.