Adding new optional arg to specify ecr region
One of the first steps of the plugin is to run ecr describe-repositories to find data of what it's to publish. Right now, it has no option but to use the default region, which leads to issues when handling multiple regions of ecr. This commit adds an argument to pass in the region to the command's call too.